README.md tiles
The goal of Piano Tiles is for the player to tap on the black tiles as they appear from the top of the screen while avoiding the white. When a black tile is tapped, a piano sound is produced. Famous compositions such as "Für Elise" and "Ode to Joy" are formed in quick succession.